separated stackable config
